better hooks api with global useXYZ() exports
