+++ package.json
@@ -6,9 +6,9 @@
   "devDependencies": {
     "@types/jest": "^24.0.18",
     "@types/node": "^12.7.4",
-    "@types/prosemirror-model": "^1.7.2",
-    "@types/prosemirror-state": "^1.2.5",
-    "@types/prosemirror-view": "^1.15.0",
+    "@types/prosemirror-model": "^1.13.2",
+    "@types/prosemirror-state": "^1.2.7",
+    "@types/prosemirror-view": "^1.19.1",
     "@typescript-eslint/eslint-plugin": "^2.19.2",
     "@typescript-eslint/parser": "^2.19.2",
     "eslint": "^6.0.1",
