+++ src/index.ts
@@ -1,3 +1,11 @@
 export * from "./plugin";
 export { getReactedDoc } from "./doc";
 export { addTemporaryIdsToDoc } from "./util";
+export {
+    useState,
+    useEffect,
+    useRef,
+    useDeferredNode,
+    useTransactionState,
+    useDocumentState,
+} from "./globalHooks";
