+++ src/store/__tests__/attrStore.test.ts
@@ -1,6 +1,7 @@
 /* global it, expect, jest */
 import { Node } from "prosemirror-model";
 
+import { useState, useEffect, useRef, useDocumentState } from "../..";
 import { createSchema, greeter } from "../../examples/schemas";
 import { AttrStore } from "../attrStore";
 
@@ -10,8 +11,8 @@ const schema = createSchema({ greeter });
 
 const testNode = Node.fromJSON(schema, { type: "greeter", attrs: { name: "world" } });
 
-const createTestStore = (fn, globalHooks = {}, onInvalidate?) =>
-    new AttrStore("anything", fn, globalHooks, onInvalidate);
+const createTestStore = (fn, documentHooks = {}, onInvalidate?) =>
+    new AttrStore("anything", fn, documentHooks as any, onInvalidate);
 
 it("runs an attr that can access the current Node value", () => {
     const attr = createTestStore(function stateCell(node) {
@@ -24,17 +25,16 @@ it("runs an attr that can access the current Node value", () => {
 it("runs an attr that can access the current global hooks", () => {
     const attr = createTestStore(
         function stateCell() {
-            return this.use3();
+            return useDocumentState(["hey"]);
         },
-        { use3: () => 3 }
+        { useDocumentState: ([val]) => `${val}? yep.` }
     );
     const result = attr.run(testNode);
-    expect(result).toEqual(3);
+    expect(result).toEqual("hey? yep.");
 });
 
 it("runs an attr with a useState call", () => {
     const attr = createTestStore(function stateCell() {
-        const { useState } = this;
         const [someState] = useState("hello!");
         return someState;
     });
@@ -44,7 +44,6 @@ it("runs an attr with a useState call", () => {
 
 it("runs an attr with a useRef call", () => {
     const attr = createTestStore(function stateCell() {
-        const { useRef } = this;
         const count = useRef(-1);
         count.current += 1;
         return count.current;
@@ -58,7 +57,6 @@ it("runs an attr with a useRef call", () => {
 
 it("runs an attr with a useState and a useEffect call", () => {
     const attr = createTestStore(function stateCell() {
-        const { useState, useEffect } = this;
         const [count, setCount] = useState(0);
 
         // In a DocumentStore this would trigger an infinite loop
@@ -75,7 +73,6 @@ it("runs an attr with a useState and a useEffect call", () => {
 
 it("runs another attr that uses both useState and useEffect", () => {
     const attr = createTestStore(function() {
-        const { useState, useEffect } = this;
         const [count, setCount] = useState(37);
 
         useEffect(() => {
@@ -98,7 +95,6 @@ it("re-runs a useEffect hook only when its dependencies change", () => {
     let b = 5;
 
     const attr = createTestStore(function() {
-        const { useEffect } = this;
         useEffect(() => {
             callbackFn();
             return teardownFn;
@@ -136,7 +132,6 @@ it("tears down a useEffect hook as expected", () => {
     let useFirstTeardown = true;
 
     const attr = createTestStore(function() {
-        const { useEffect } = this;
         useEffect(() => {
             return useFirstTeardown ? firstTeardown : secondTeardown;
         });
@@ -164,7 +159,6 @@ it("invalidates when its state changes", () => {
 
     const attr = createTestStore(
         function() {
-            const { useEffect, useState } = this;
             const [ready, setReady] = useState(false);
 
             useEffect(() => {
@@ -187,7 +181,6 @@ it("does not invalidate when setState is called, but does not change the state",
 
     const attr = createTestStore(
         function() {
-            const { useEffect, useState } = this;
             const [ready, setReady] = useState(false);
 
             useEffect(() => {
@@ -210,7 +203,6 @@ it("does not invalidate after it has been destroyed", () => {
 
     const attr = createTestStore(
         function() {
-            const { useEffect, useState } = this;
             const [ready, setReady] = useState(false);
 
             useEffect(() => {
