+++ src/store/util.ts
@@ -10,12 +10,6 @@ export const throwError = (message: string) => {
     throw new Error(`${prefix} ${message}`);
 };
 
-const isValidReactiveFn = fn => {
-    // Arrow functions have no prototype, and we need non-arrows so we can bind `this` to them.
-    // eslint-disable-next-line no-prototype-builtins
-    return typeof fn === "function" && fn.hasOwnProperty("prototype");
-};
-
 export const assertCanCreateReactiveNodeStore = (
     nodeType: string,
     spec: ReactiveNodeSpec,
@@ -26,14 +20,6 @@ export const assertCanCreateReactiveNodeStore = (
     if (hasIdAttr) {
         const hasReactiveAttrs = reactiveAttrs && Object.keys(reactiveAttrs).length > 0;
         if (hasReactiveAttrs) {
-            const attrWithWrongType = Object.entries(reactiveAttrs).find(
-                entry => !isValidReactiveFn(entry[1])
-            );
-            if (attrWithWrongType) {
-                throwError(
-                    `Reactive attr ${attrWithWrongType} on ${nodeType} must be a (non-arrow) function`
-                );
-            }
             const reactiveAttrShadowingAttr = Object.keys(reactiveAttrs).find(
                 attr => !!attrs[attr]
             );
